Remove false positive B608 report
Codacy reported "Possible SQL injection vector through string-based query construction".  This was a false positive, as queries were built with a hardcoded table const being added so that the table could be referenced elsewhere.  Fixed anyway, for a clean bill of health.
